Move days list into schedule days
