+++ src/declaration.d.ts
@@ -1,1 +1,5 @@
 declare module '*.png';
+declare module '*.jpg';
+declare module '*.jpeg';
+declare module '*.gif';
+declare module '*.svg';
